refactor: 켄벤션에 맞도록 api 호출 수정, 컴포넌트 이름 수정
