+++ src/context.ts
@@ -1,9 +1,11 @@
 import { PostAPI, PostAPIInterface } from './data/postAPI'
 import AccountAPI, { AccountAPIInterface } from './data/accountAPI'
+import ChatAPI, { ChatAPIInterface } from './data/chatAPI'
 
 export interface ContextInterface {
     postAPI: PostAPIInterface
     signUpAPI: AccountAPIInterface
+    chatAPI: ChatAPIInterface
 }
 
 export class Context implements ContextInterface {
@@ -11,8 +13,11 @@ export class Context implements ContextInterface {
 
     signUpAPI
 
+    chatAPI
+
     constructor() {
         this.postAPI = new PostAPI()
         this.signUpAPI = new AccountAPI()
+        this.chatAPI = new ChatAPI()
     }
 }
