--- src/presentation/components/launcher/LauncherChat.tsx
@@ -1,3 +1,0 @@
-export default function LauncherChat() {
-    return <span />
-}
