--- src/presentation/components/launcher/LauncherLayout.tsx
@@ -1,3 +1,0 @@
-export default function LauncherLayout() {
-    return <>foo</>
-}
