+++ src/presentation/components/launcher/LauncherMessageComponent.tsx
@@ -2,20 +2,20 @@ import React from 'react'
 import styled from 'styled-components'
 import { ChatWithUser } from '../../../types/chat'
 
-export const MessageWrapper = styled.div<{ me: boolean }>`
+const MessageWrapper = styled.div<{ me: boolean }>`
     position: relative;
     display: flex;
     padding: 0.5em 0;
     flex-direction: ${(props) => (props.me ? 'row-reverse' : 'row')};
 `
 
-export const MessageInfo = styled.div`
+const MessageInfo = styled.div`
     display: flex;
     flex-direction: column;
     padding: 0 10px;
     align-items: flex-start;
 `
-export const Name = styled.span`
+const Name = styled.span`
     display: inline-block;
     font-weight: 600;
     height: 40px;
@@ -23,7 +23,7 @@ export const Name = styled.span`
     color: rgba(0, 0, 0, 0.75);
 `
 
-export const Message = styled.p`
+const Message = styled.p`
     margin: 0;
     padding: 0;
     font-size: 0.95rem;
@@ -34,7 +34,7 @@ interface Props extends ChatWithUser {
     me: boolean
 }
 
-function ChatMessage({ id, user, message, me }: Props) {
+function LauncherMessageComponent({ id, user, message, me }: Props) {
     return (
         <MessageWrapper key={id} me={me}>
             <img src={user.image} alt="" width={30} height={30} />
@@ -50,4 +50,4 @@ function ChatMessage({ id, user, message, me }: Props) {
     )
 }
 
-export default ChatMessage
+export default LauncherMessageComponent
