+++ src/presentation/components/launcher/LauncherWriteComponent.tsx
@@ -1,5 +1,36 @@
+import styled from 'styled-components'
 import { useForm } from 'react-hook-form'
-import { ActionBox, MessageArea, WriteForm } from './index.style'
+
+const WriteForm = styled.form`
+    width: 100%; // calc(100% - 50px);
+    height: 120px;
+    display: flex;
+    flex-direction: column;
+    border-radius: 10px;
+`
+
+const MessageArea = styled.textarea`
+    resize: none;
+    height: 100px;
+    padding: 10px;
+    margin: 5px 10px;
+    outline: none;
+    border: none;
+    border-bottom: 1px solid var(--gray-100);
+    border-radius: 10px 10px;
+`
+
+const ActionBox = styled.button`
+    width: 60px;
+    padding: 5px;
+    margin: 5px 10px;
+    border-radius: 10px;
+    transition: color 200ms ease-out;
+    cursor: pointer;
+    &:hover {
+        color: var(--primary);
+    }
+`
 
 export interface MessageData {
     message: string
@@ -13,7 +44,7 @@ interface WriteProps {
     onSendMessage: (messageData: MessageData) => void
 }
 
-export default function Write({ onSendMessage }: WriteProps) {
+export default function LauncherWriteComponent({ onSendMessage }: WriteProps) {
     const { register, handleSubmit, reset } = useForm<WriteFormData>()
 
     const onSubmit = async (data: WriteFormData) => {
